fix: rating 자료형 double로 변환

+++ src/main/java/com/alphano/alphano/domain/userRating/dao/UserRatingQueryRepository.java
@@ -62,7 +62,7 @@ public class UserRatingQueryRepository {
         return PageableExecutionUtils.getPage(content, pageable, countQuery::fetchOne);
     }
 
-    public Integer findCurrentRating(Long problemId, Long userId) {
+    public Double findCurrentRating(Long problemId, Long userId) {
         QUserRating userRating = QUserRating.userRating;
 
         return queryFactory
